Fixed a bug in firmware in lightout.
There's still a bug which I believe is currently believe is firmware relation that is driving me crazy:

Interrupted fades are buggy. Here's one reproducible symptom that may hold a clue as to the larger problem:

I set up the following two threads:
 1) on play [fadeallto green] 2) on play [wait 4 fadealltored]
When I run it from the lightoff state (either on power up or by executing a lightoff after it's run) the
cross fades are starnge. But if I run it after a fadeout, the cross fades work normally.

That's really puzzling since nominally lightoff and fadeout should end up in the exact same state.