Fixed fade interrupt bug!
The problem was that, previously, during a fade update the intermediate RGBW vals
were only being updated on a particular RGBW channel if that particular channel
was changing values during the fade. This leads to the possiblility that the
int RGBW values (which are captured at the the start of an interupted fade)
might not accurately reflect the current RGBW state. (For example, consider the following two treads:

     on play [fadeto green]
     on play [wait 3 fadeto red]

This executes properly the first time. So the red channel value will be at 4095 and intred will
also be 4095. But then suppose you execute a lightout. This changes the red channel to zero, but int red will still be
4095. This will screw up the next play.

So the problem occurs due to a change in RGBW val that is not properly tracked by the the int. values.

The fix was to add updates to the int values in lightoff, and also in setcolor.